re-pull
Wrongly merged into branch patch-5, re-pull for merging into main. Initiate to divide all function into several files, each file name states purpose of grouped functions
