+++ materials/dicts.json
@@ -3,9 +3,7 @@
           "Pelabuhan Asal","Tanggal Tiba","Posisi Tolak","Pelabuhan Tujuan","Tanggal Tolak"],
   "FLKK":["SPB","Komoditi B","Ton B","M3 B","Unit B","Orang B","Komoditi M","Ton M","M3 M","Unit M","Orang M"],
   "COMBINED_SET":["SPB","Tipe Kapal","Kapal","GT","Nakhoda","Call Sign","Jumlah Awak","Bendera","Agen",
-                  "Pelabuhan Asal","Tanggal Tiba","Komoditi B","Ton B","M3 B","Unit B","Orang B",
-                  "Posisi Tolak","Pelabuhan Tujuan","Tanggal Tolak","Komoditi M","Ton M","M3 M",
-                  "Unit M","Orang M"],
+                  "Pelabuhan Asal","Tanggal Tiba","Komoditi B","Ton B","M3 B","Unit B","Orang B","Posisi Tolak","Pelabuhan Tujuan","Tanggal Tolak","Komoditi M","Ton M","M3 M","Unit M","Orang M"],
   "OUTPUT_HEADER":["NO","SPB TOLAK","PPK 27","PPK 29","NAMA KAPAL","GT","NT","TANDA SELAR",
                    "TEMPAT PENDAFTARAN","NAKHODA","TANDA PANGGIL","CREW","BENDERA","PEMILIK/AGEN","ASAL",
                    "TANGGAL TIBA","JAM TIBA","TIBA ISI/KOSONG","MUATAN BONGKAR","JUMLAH MUATAN BONGKAR",
